# apps/api/src/main/java/com/growtogether/auth/dto/SignUpRequest.java
package com.growtogether.auth.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record SignUpRequest(
    @NotBlank
    @Size(max = 50)
    String name,

    @NotBlank
    @Email
    @Size(max = 254)
    String email,

    @NotBlank
    @Size(min = 8, max = 72)
    String password
) {
    @Override public String toString() { return "SignUpRequest[redacted]"; }
}